test: fixed tests
