+++ app/lib/snaps/SnapsExecutionWebView.test.tsx
@@ -6,7 +6,6 @@ import {
   removeWebView,
   SnapsExecutionWebView,
 } from './SnapsExecutionWebView';
-import { buildE2EProxyPatchScript } from './SnapsE2EProxy';
 
 describe('SnapsExecutionWebView', () => {
   it('should render correctly', () => {
@@ -34,39 +33,5 @@ describe('SnapsExecutionWebView', () => {
     wrapper.rerender(<SnapsExecutionWebView />);
     expect(await wrapper.queryByTestId('foo')).toBeNull();
     expect(await wrapper.queryByTestId('bar')).toBeTruthy();
-  });
-
-  it('builds an iOS proxy patch script without emulator host fallback', () => {
-    const script = buildE2EProxyPatchScript({
-      mockServerPort: '8000',
-      platform: 'ios',
-      snapId: 'npm:@metamask/solana-wallet-snap',
-    });
-
-    expect(script).toContain('/proxy?source=');
-    expect(script).toContain("snapProxySource = 'snap-webview'");
-    expect(script).toContain('snapId = "npm:@metamask/solana-wallet-snap"');
-    expect(script).toContain(
-      '?source=${snapProxySource}&snapId=${encodeURIComponent(snapId)}&url=',
-    );
-    expect(script).toContain('http://localhost');
-    expect(script).not.toContain('10.0.2.2');
-  });
-
-  it('builds an Android proxy patch script with emulator host fallback', () => {
-    const script = buildE2EProxyPatchScript({
-      mockServerPort: '8000',
-      platform: 'android',
-      snapId: 'npm:@metamask/bitcoin-wallet-snap',
-    });
-
-    expect(script).toContain('/proxy?source=');
-    expect(script).toContain("snapProxySource = 'snap-webview'");
-    expect(script).toContain('snapId = "npm:@metamask/bitcoin-wallet-snap"');
-    expect(script).toContain(
-      '?source=${snapProxySource}&snapId=${encodeURIComponent(snapId)}&url=',
-    );
-    expect(script).toContain('http://localhost');
-    expect(script).toContain('10.0.2.2');
   });
 });
